requirements.txt
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
segmentation-models-pytorch>=0.3.0
albumentations>=1.3.0
opencv-python>=4.7.0
numpy>=1.22.0
Pillow>=9.4.0
PyYAML>=6.0
tqdm>=4.65.0
matplotlib>=3.7.0
pandas>=1.5.0
seaborn>=0.12.0
scikit-learn>=1.2.0 